==> 3_Scraping More Informartion.ipynb <==
# 1.0 Import Dependencies

import os, platform, time, urllib.request, openpyxl, operator
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.common.action_chains import ActionChains
from openpyxl import Workbook
from urllib.request import urlopen
import sys, requests, re, json
from pandas.io.json import json_normalize
from selenium.webdriver.common.keys import Keys
from bs4 import BeautifulSoup

import pandas as pd
pd.set_option('display.max_columns', None)
import numpy as np

import datetime as dt

import warnings
warnings.filterwarnings('ignore')

import utils
from utils import get_information

# 2.0 Load Data Influencer Names

inf_names = pd.read_csv(r'influencer_post_links.csv')

# 3.0 Get Information
- 10000 Rows
- It takes 3 Hours to run this code per batch
- I create to take information by batching to efficient time

## 3.1 Batch 1

batch_1 = inf_names[inf_names.index <= 10000]

#iterate every post links and get the information
for link in batch_1['link']:
    likes, comment_counts, dates, captions, type_posts, links = get_information(link)

data_batch_1 = pd.DataFrame(links).drop_duplicates().reset_index(drop=True).rename(columns={
    0:'link'
})

data_batch_1['likes'] = likes
data_batch_1['comment_counts'] = comment_counts
data_batch_1['dates'] = dates
data_batch_1['captions'] = captions
data_batch_1['type_posts'] = type_posts

data_batch_1_2 = pd.merge(data_batch_1, batch_1, on=['link'], how='left').reset_index(drop=True)

# data_batch_1_2.to_csv('data_batch_1.csv', index=False)

#Clear the list so data in this batch and the next batch are not overlapped
likes.clear(), comment_counts.clear(), dates.clear(), captions.clear(), type_posts.clear(), links.clear()

## 3.2 Batch 2

batch_2 = inf_names[(inf_names.index > 10000) & (inf_names.index <= 20000)].reset_index(drop=True)

#iterate every post links and get the information
for link in batch_2['link']:
    likes, comment_counts, dates, captions, type_posts, links = get_information(link)

data_batch_2 = pd.DataFrame(links).drop_duplicates().reset_index(drop=True).rename(columns={
    0:'link'
})

data_batch_2['likes'] = likes
data_batch_2['comment_counts'] = comment_counts
data_batch_2['dates'] = dates
data_batch_2['captions'] = captions
data_batch_2['type_posts'] = type_posts

data_batch_2_2 = pd.merge(data_batch_2, batch_2, on=['link'], how='left').reset_index(drop=True)

data_batch_2_3 = data_batch_2_2[data_batch_2_2['username'].notnull()].reset_index(drop=True)

# data_batch_2_3.to_csv('data_batch_2.csv', index=False)

#Clear the list so data in this batch and the next batch are not overlapped
likes.clear(), comment_counts.clear(), dates.clear(), captions.clear(), type_posts.clear(), links.clear()

## 3.3 Batch 3

batch_3 = inf_names[(inf_names.index > 20000) & (inf_names.index <= 30000)].reset_index(drop=True)

#iterate every post links and get the information
for link in batch_3['link']:
    likes, comment_counts, dates, captions, type_posts, links = get_information(link)

data_batch_3 = pd.DataFrame(links).drop_duplicates().reset_index(drop=True).rename(columns={
    0:'link'
})

data_batch_3['likes'] = likes
data_batch_3['comment_counts'] = comment_counts
data_batch_3['dates'] = dates
data_batch_3['captions'] = captions
data_batch_3['type_posts'] = type_posts

data_batch_3_2 = pd.merge(data_batch_3, batch_3, on=['link'], how='left').reset_index(drop=True)

data_batch_3_3 = data_batch_3_2[data_batch_3_2['username'].notnull()].reset_index(drop=True)

# data_batch_3_3.to_csv('data_batch_3.csv', index=False)

#Clear the list so data in this batch and the next batch are not overlapped
likes.clear(), comment_counts.clear(), dates.clear(), captions.clear(), type_posts.clear(), links.clear()

## 3.4 Batch 4

batch_4 = inf_names[(inf_names.index > 30000) & (inf_names.index <= 40000)].reset_index(drop=True)

#iterate every post links and get the information
for link in batch_4['link']:
    likes, comment_counts, dates, captions, type_posts, links = get_information(link)

data_batch_4 = pd.DataFrame(links).drop_duplicates().reset_index(drop=True).rename(columns={
    0:'link'
})

data_batch_4['likes'] = likes
data_batch_4['comment_counts'] = comment_counts
data_batch_4['dates'] = dates
data_batch_4['captions'] = captions
data_batch_4['type_posts'] = type_posts

data_batch_4_2 = pd.merge(data_batch_4, batch_4, on=['link'], how='left').reset_index(drop=True)

data_batch_4_3 = data_batch_4_2[data_batch_4_2['username'].notnull()].reset_index(drop=True)

# data_batch_4_3.to_csv('data_batch_4.csv', index=False)

#Clear the list so data in this batch and the next batch are not overlapped
likes.clear(), comment_counts.clear(), dates.clear(), captions.clear(), type_posts.clear(), links.clear()

## 3.5 Batch 5

batch_5 = inf_names[(inf_names.index > 40000) & (inf_names.index <= 50000)].reset_index(drop=True)

#iterate every post links and get the information
for link in batch_5['link']:
    likes, comment_counts, dates, captions, type_posts, links = get_information(link)

data_batch_5 = pd.DataFrame(links).drop_duplicates().reset_index(drop=True).rename(columns={
    0:'link'
})

data_batch_5['likes'] = likes
data_batch_5['comment_counts'] = comment_counts
data_batch_5['dates'] = dates
data_batch_5['captions'] = captions
data_batch_5['type_posts'] = type_posts

data_batch_5_2 = pd.merge(data_batch_5, batch_5, on=['link'], how='left').reset_index(drop=True)

data_batch_5_3 = data_batch_5_2[data_batch_5_2['username'].notnull()].reset_index(drop=True)

# data_batch_5_3.to_csv('data_batch_5.csv', index=False)

#Clear the list so data in this batch and the next batch are not overlapped
likes.clear(), comment_counts.clear(), dates.clear(), captions.clear(), type_posts.clear(), links.clear()

## 3.6 Batch 6

batch_6 = inf_names[(inf_names.index > 50000)].reset_index(drop=True)

batch_6

#iterate every post links and get the information
for link in batch_6['link']:
    likes, comment_counts, dates, captions, type_posts, links = get_information(link)

data_batch_6 = pd.DataFrame(links).drop_duplicates().reset_index(drop=True).rename(columns={
    0:'link'
})

data_batch_6['likes'] = likes
data_batch_6['comment_counts'] = comment_counts
data_batch_6['dates'] = dates
data_batch_6['captions'] = captions
data_batch_6['type_posts'] = type_posts

data_batch_6_2 = pd.merge(data_batch_6, batch_6, on=['link'], how='left').reset_index(drop=True)

data_batch_6_3 = data_batch_6_2[data_batch_6_2['username'].notnull()].reset_index(drop=True)

data_batch_6_3.to_csv('data_batch_6.csv', index=False)

#Clear the list so data in this batch and the next batch are not overlapped
likes.clear(), comment_counts.clear(), dates.clear(), captions.clear(), type_posts.clear(), links.clear()
