==> 4_Data Preprocessing.ipynb <==
# 1.0 Import Dependencies

import pandas as pd
pd.set_option('display.max_columns', None)
import numpy as np

import datetime as dt

import seaborn as sns
import matplotlib.pyplot as plt

import warnings
warnings.filterwarnings('ignore')

# 2.0 Load Data

#Load all data from bacth 1 untul 5
data_batch_1 = pd.read_csv(r'data_batch_1.csv')
data_batch_2 = pd.read_csv(r'data_batch_2.csv')
data_batch_3 = pd.read_csv(r'data_batch_3.csv')
data_batch_4 = pd.read_csv(r'data_batch_4.csv')
data_batch_5 = pd.read_csv(r'data_batch_5.csv')
data_batch_6 = pd.read_csv(r'data_batch_6.csv')

#general information data
data_general = pd.read_csv(r'influencer_names_final.csv')

#concat all data
data_detail = pd.concat([data_batch_1,
           data_batch_2,
           data_batch_3,
           data_batch_4,
           data_batch_5,
           data_batch_6],
          ignore_index=True)

#join with general data
df = pd.merge(data_detail, data_general, how='left')

#top 5 our data
df.head()

#Initially i have 65.998 rows and 10 features
df.shape

# 3.0 Data Preprocessing Stage

## 3.1 Convert Feature Dates
- because the dates format is in epochs, so we must convert it to general datetime format
- the new format is year-month-day-hour when the post is being posted by influencers

df['dates'] = df['dates'].apply(lambda x: dt.datetime.fromtimestamp(x).strftime('%Y-%m-%d-%H'))

print(df['dates'].min())
print(df['dates'].max())

## 3.2 Cleansing Features Captions

df['captions'] = df['captions'].replace(r'[\n]', '', regex=True)

df.isnull().sum()

#fill missing value in features captions
df['captions'] = df['captions'].fillna('No Captions')

# 4.0 Feature Engineering Stage 1
- this is my own definition related to engagement.
    - likes_engagement = number of likes each post / number of followers
    - comments_engagement = number of comments each post / number of followers
    - total_engagement = likes_engagement + comments_engagement
    - our objective is to predict total engagemnt

## 4.1 Feature dates

#separte for future use
df['year'] = df['dates'].apply(lambda x: x[:4])
df['month'] = df['dates'].apply(lambda x: x[5:7])
df['day'] = df['dates'].apply(lambda x: x[8:10])
df['hour'] = df['dates'].apply(lambda x: x[11:])

#create year_month features
df['year_month'] = df['year'] + df['month']

#because the count of post before 2018 is small, we only filter post that >= 2018
#and our objective is to predict avg engagment rate in July 2020, so we must exclude July 2020 too.
df_2 = df[df['year']>='2018'].reset_index(drop=True)
df_3 = df_2[df_2['year_month']!='202007'].reset_index(drop=True)

## 4.2 Feature Engagement

df_3['likes_engagement'] = df_3['likes'] / df_3['followers'] * 100
df_3['comments_engagement'] = df_3['comment_counts'] / df_3['followers'] * 100

df_3['total_engagement'] = df_3['likes_engagement'] + df_3['comments_engagement']

## 4.3 Feature Captions

#make lower case
df_3['captions'] = df_3['captions'].str.lower()

#delete whitespace
df_3['captions'] = df_3['captions'].str.strip()

#feature length of captions
df_3['len_capt'] = df_3['captions'].apply(lambda x: len(x))

#create feature length of captions without punctuation
df_3['len_capt_wo_punct'] = df_3['captions'].str.replace(r'[^\w\s]', '', regex=True).apply(lambda x: len(x))

#feature length of punctuation only
df_3['len_capt_punct'] = df_3['captions'].str.replace(r'[\w\s]', '', regex=True).apply(lambda x: len(x))

#feature number of words in each captions
df_3['n_words'] = df_3['captions'].apply(lambda x: len(str(x).split(" ")))

#average char per words
df_3['avg_char_words'] = df_3['n_words'] / df_3['len_capt']

#number of numeric char in captions
df_3['n_numeric'] = df_3['captions'].apply(lambda x: len([x for x in x.split() if x.isdigit()]))

df_3.head()

#from our stage 1 preprocessing and feature engineering, we can create 14 new features.
df_3.shape

df_3.to_csv('data_preprocessing_1.csv', index=False)
