==> 5_Feature Engineering.ipynb <==
# 1.0 Import Dependencies

import pandas as pd
pd.set_option('display.max_columns', None)
import numpy as np

import datetime as dt

import seaborn as sns
import matplotlib.pyplot as plt

import warnings
warnings.filterwarnings('ignore')

# 2.0 Load Data

df = pd.read_csv(r'data_preprocessing_1.csv')

df.head()

# 3.0 Create Base Table

base_table = df.groupby(['username', 'year_month']).agg({
    'link' : 'count'
}).reset_index().rename(columns={
    'link' : 'n_post'
})

#create features lag of n_post (last 3 month)
#number of n_post 1 months ago
base_table['n_post_01'] = base_table.groupby(['username'])['n_post'].shift(1).fillna(0)

#number of n_post 2 months ago
base_table['n_post_02'] = base_table.groupby(['username'])['n_post'].shift(2).fillna(0)

#number of n_post 3 months ago
base_table['n_post_03'] = base_table.groupby(['username'])['n_post'].shift(3).fillna(0)

## 3.1 Features type_posts

df_type_post = df.groupby(['username', 'year_month'])['type_posts'].value_counts().unstack(2).reset_index().fillna(0)

base_table['n_img_post'] = df_type_post['GraphImage']
base_table['n_vid_post'] = df_type_post['GraphVideo']
base_table['n_sidecar_post'] = df_type_post['GraphSidecar']

## 3.2 Features Likes

#create features sum, avg and stdev of likes
base_table['sum_likes'] = df.groupby(['username', 'year_month'])['likes'].sum().reset_index()['likes']
base_table['avg_likes'] = df.groupby(['username', 'year_month'])['likes'].mean().reset_index()['likes']
base_table['std_likes'] = df.groupby(['username', 'year_month'])['likes'].std().reset_index()['likes'].fillna(0)

#create features lag of sum_likes and mean_likes (last 3 month)
base_table['avg_likes_01'] = base_table.groupby(['username'])['avg_likes'].shift(1).fillna(0)
base_table['avg_likes_02'] = base_table.groupby(['username'])['avg_likes'].shift(2).fillna(0)
base_table['avg_likes_03'] = base_table.groupby(['username'])['avg_likes'].shift(3).fillna(0)

base_table['sum_likes_01'] = base_table.groupby(['username'])['sum_likes'].shift(1).fillna(0)
base_table['sum_likes_02'] = base_table.groupby(['username'])['sum_likes'].shift(2).fillna(0)
base_table['sum_likes_03'] = base_table.groupby(['username'])['sum_likes'].shift(3).fillna(0)

## 3.3 Features Comments

#create features sum, avg and stdev of comments
base_table['sum_comments'] = df.groupby(['username', 'year_month'])['comment_counts'].sum().reset_index()['comment_counts']
base_table['avg_comments'] = df.groupby(['username', 'year_month'])['comment_counts'].mean().reset_index()['comment_counts']
base_table['std_comments'] = df.groupby(['username', 'year_month'])['comment_counts'].std().reset_index()['comment_counts'].fillna(0)

#create features lag of sum_comments and mean_comments (last 3 month)
base_table['avg_comments_01'] = base_table.groupby(['username'])['avg_comments'].shift(1).fillna(0)
base_table['avg_comments_02'] = base_table.groupby(['username'])['avg_comments'].shift(2).fillna(0)
base_table['avg_comments_03'] = base_table.groupby(['username'])['avg_comments'].shift(3).fillna(0)

base_table['sum_comments_01'] = base_table.groupby(['username'])['sum_comments'].shift(1).fillna(0)
base_table['sum_comments_02'] = base_table.groupby(['username'])['sum_comments'].shift(2).fillna(0)
base_table['sum_comments_03'] = base_table.groupby(['username'])['sum_comments'].shift(3).fillna(0)

## 3.4 Features Captions

#create features sum, avg and stdev of captions_wo_punct
base_table['sum_capt_wo_punct'] = df.groupby(['username', 'year_month'])['len_capt_wo_punct'].sum().reset_index()['len_capt_wo_punct']
base_table['avg_capt_wo_punct'] = df.groupby(['username', 'year_month'])['len_capt_wo_punct'].mean().reset_index()['len_capt_wo_punct']
base_table['std_capt_wo_punct'] = df.groupby(['username', 'year_month'])['len_capt_wo_punct'].std().reset_index()['len_capt_wo_punct'].fillna(0)

#create features sum, avg and stdev of captions_w_punct
base_table['sum_capt_w_punct'] = df.groupby(['username', 'year_month'])['len_capt_punct'].sum().reset_index()['len_capt_punct']
base_table['avg_capt_w_punct'] = df.groupby(['username', 'year_month'])['len_capt_punct'].mean().reset_index()['len_capt_punct']
base_table['std_capt_w_punct'] = df.groupby(['username', 'year_month'])['len_capt_punct'].std().reset_index()['len_capt_punct'].fillna(0)

#create features sum, avg and stdev of n_words
base_table['sum_n_words'] = df.groupby(['username', 'year_month'])['n_words'].sum().reset_index()['n_words']
base_table['avg_n_words'] = df.groupby(['username', 'year_month'])['n_words'].mean().reset_index()['n_words']
base_table['std_n_words'] = df.groupby(['username', 'year_month'])['n_words'].std().reset_index()['n_words'].fillna(0)

#create features sum, avg and stdev of avg_char_words
base_table['sum_avg_char_words'] = df.groupby(['username', 'year_month'])['avg_char_words'].sum().reset_index()['avg_char_words']
base_table['avg_avg_char_words'] = df.groupby(['username', 'year_month'])['avg_char_words'].mean().reset_index()['avg_char_words']
base_table['std_avg_char_words'] = df.groupby(['username', 'year_month'])['avg_char_words'].std().reset_index()['avg_char_words'].fillna(0)

#create features lag of captions (last 3 month)
base_table['sum_capt_wo_punct_01'] = base_table.groupby(['username'])['sum_capt_wo_punct'].shift(1).fillna(0)
base_table['sum_capt_wo_punct_02'] = base_table.groupby(['username'])['sum_capt_wo_punct'].shift(2).fillna(0)
base_table['sum_capt_wo_punct_03'] = base_table.groupby(['username'])['sum_capt_wo_punct'].shift(3).fillna(0)

base_table['sum_capt_w_punct_01'] = base_table.groupby(['username'])['sum_capt_w_punct'].shift(1).fillna(0)
base_table['sum_capt_w_punct_02'] = base_table.groupby(['username'])['sum_capt_w_punct'].shift(2).fillna(0)
base_table['sum_capt_w_punct_03'] = base_table.groupby(['username'])['sum_capt_w_punct'].shift(3).fillna(0)

base_table['sum_n_words_01'] = base_table.groupby(['username'])['sum_n_words'].shift(1).fillna(0)
base_table['sum_n_words_02'] = base_table.groupby(['username'])['sum_n_words'].shift(2).fillna(0)
base_table['sum_n_words_03'] = base_table.groupby(['username'])['sum_n_words'].shift(3).fillna(0)

base_table['sum_avg_char_words_01'] = base_table.groupby(['username'])['sum_avg_char_words'].shift(1).fillna(0)
base_table['sum_avg_char_words_02'] = base_table.groupby(['username'])['sum_avg_char_words'].shift(2).fillna(0)
base_table['sum_avg_char_words_03'] = base_table.groupby(['username'])['sum_avg_char_words'].shift(3).fillna(0)

base_table['sum_n_numeric'] = df.groupby(['username', 'year_month'])['n_numeric'].sum().reset_index()['n_numeric']

## 3.5 Features Engagement

#create features sum and avg engagement
base_table['sum_likes_engagement'] = df.groupby(['username', 'year_month'])['likes_engagement'].sum().reset_index()['likes_engagement']
base_table['sum_comments_engagement'] = df.groupby(['username', 'year_month'])['comments_engagement'].sum().reset_index()['comments_engagement']
base_table['sum_total_engagement'] = df.groupby(['username', 'year_month'])['total_engagement'].sum().reset_index()['total_engagement']

base_table['avg_likes_engagement'] = df.groupby(['username', 'year_month'])['likes_engagement'].mean().reset_index()['likes_engagement']
base_table['avg_comments_engagement'] = df.groupby(['username', 'year_month'])['comments_engagement'].mean().reset_index()['comments_engagement']
base_table['avg_total_engagement'] = df.groupby(['username', 'year_month'])['total_engagement'].mean().reset_index()['total_engagement']

#create features lag of engagement (last 3 month)
base_table['avg_total_engagement_01'] = base_table.groupby(['username'])['avg_total_engagement'].shift(1).fillna(0)
base_table['avg_total_engagement_02'] = base_table.groupby(['username'])['avg_total_engagement'].shift(2).fillna(0)
base_table['avg_total_engagement_03'] = base_table.groupby(['username'])['avg_total_engagement'].shift(3).fillna(0)

## 3.6 Target Variable

#Create our target Variable
base_table['y_month_01'] = base_table.groupby(['username'])['avg_total_engagement'].shift(-1).fillna(0)

base_table

base_table.to_csv('data_feature_engineering.csv', index=False)
